[service] The storage service sets up the LSM
